third_party/blink/web_tests/platform/generic/fast/frames/frame-element-name-expected.txt



--------
Frame: 'left'
--------
PASS escape(window.frameElement.name) is "left"
PASS escape(window.name) is "left"
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS escape(window.frameElement.name) is ""
PASS escape(window.name) is ""
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
PASS escape(window.frameElement.name) is "_blank"
PASS escape(window.name) is "_blank"
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '_parent'
--------
PASS escape(window.frameElement.name) is "_parent"
PASS escape(window.name) is "_parent"
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '_self'
--------
PASS escape(window.frameElement.name) is "_self"
PASS escape(window.name) is "_self"
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '_top'
--------
PASS escape(window.frameElement.name) is "_top"
PASS escape(window.name) is "_top"
PASS successfullyParsed is true

TEST COMPLETE
